# Reduce the memory impact