--- a/FAForever.Replay/CommandData.cs
+++ b/FAForever.Replay/CommandData.cs
@@ -1,5 +1,5 @@
 ﻿
 namespace FAForever.Replay
 {
-    public record struct CommandData(int Identifier, CommandType Type, CommandTarget Target, CommandFormation Formation, String BlueprintId, LuaData LuaParameters, Boolean AddToQueue, byte[] Unknown1, byte[] Unknown2, byte[] Unknown3, byte[] Unknown4);
+    public record struct CommandData(int Identifier, CommandType Type, CommandTarget Target, CommandFormation Formation, String BlueprintId, LuaData LuaParameters, Boolean AddToQueue, int Unknown1, int Unknown2, byte Unknown3, int Unknown4, int Unknown5, int Unknown6);
 }