Removes the intermediate token and various other byte arrays to drastically reduce the number of allocated memory on the heap. The memory allocations are profiled using the built-in profiler of Visual Studio 2022 and via BenchmarkDotNet. See also the reference section in the readme.

Before:

| ReplayFile           | Mean      | Error     | StdDev   | Gen0      | Gen1      | Gen2      | Allocated |
|--------------------- |----------:|----------:|---------:|----------:|----------:|----------:|----------:|
| 23225104.fafreplay   | 168.52 ms | 25.965 ms | 1.423 ms | 4000.0000 | 3666.6667 | 1333.3333 | 158.37 MB |
| 23225323.fafreplay   |  61.37 ms | 25.634 ms | 1.405 ms | 2500.0000 | 2375.0000 | 1250.0000 |  78.53 MB |
| 23225440.fafreplay   |  17.21 ms |  4.253 ms | 0.233 ms | 1625.0000 | 1593.7500 |  968.7500 |  37.03 MB |
| 23225508.fafreplay   |  12.10 ms | 11.488 ms | 0.630 ms | 1250.0000 | 1234.3750 |  796.8750 |  25.46 MB |
| 23225685.fafreplay   |  64.36 ms | 72.872 ms | 3.994 ms | 2500.0000 | 2375.0000 | 1125.0000 |   79.8 MB |
| TestC(...)eplay [24] |  58.77 ms | 33.523 ms | 1.837 ms | 2000.0000 | 1888.8889 | 1000.0000 |   59.9 MB |

After: 

| ReplayFile           | Mean      | Error      | StdDev    | Gen0      | Gen1      | Gen2      | Allocated |
|--------------------- |----------:|-----------:|----------:|----------:|----------:|----------:|----------:|
| 23225104.fafreplay   | 53.056 ms | 11.4210 ms | 0.6260 ms | 1909.0909 | 1818.1818 | 1272.7273 |  56.88 MB |
| 23225323.fafreplay   | 14.862 ms | 30.0466 ms | 1.6470 ms | 1000.0000 |  968.7500 |  781.2500 |  23.13 MB |
| 23225440.fafreplay   |  6.382 ms |  0.9361 ms | 0.0513 ms | 1109.3750 | 1101.5625 |  992.1875 |   9.51 MB |
| 23225508.fafreplay   |  3.282 ms |  0.9536 ms | 0.0523 ms |  703.1250 |  699.2188 |  644.5313 |   5.56 MB |
| 23225685.fafreplay   | 19.225 ms | 30.5515 ms | 1.6746 ms |  937.5000 |  906.2500 |  687.5000 |  25.16 MB |
| TestC(...)eplay [24] | 14.766 ms | 26.3337 ms | 1.4434 ms |  718.7500 |  687.5000 |  500.0000 |  19.49 MB |